requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
distro==1.5.0
dj-database-url==0.5.0
Django==3.0.5
django-heroku==0.3.1
gunicorn==20.0.4
idna==2.9
numpy==1.18.2
pandas==1.0.3
pgeocode==0.2.1
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2019.3
pyzipcode==2.0.0
requests==2.23.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
tabula-py==2.1.0
urllib3==1.25.8
whitenoise==5.0.1
